refactor: atualiza as constantes
