+++ src/data/permissions.js
@@ -1,0 +1,5 @@
+export const PERMISSION_LEVEL = '2';
+
+export const ADMIN_MASTER = 'admin master';
+
+export const ADMIN_VOLUNTEER = 'admin volunteer';
