--- src/data/pixKey.js
@@ -1,1 +1,0 @@
-export const PERMISSION_LEVEL = '2';
